move filter types inside envoyinit project
